Extract core and pause/resume functionalities from EmsCommon
