+++ app/controllers/mixins/ems_common.rb
@@ -1,7 +1,11 @@
 module Mixins
   module EmsCommon
     extend ActiveSupport::Concern
+
+    # This is the list of extracted parts that can be used separately
+    include Core
     include Metrics
+    include PauseResume
 
     included do
       include Mixins::GenericSessionMixin
@@ -612,69 +616,6 @@ module Mixins
        [_('Non-SSL'), 'non-ssl']]
     end
 
-    def process_emss(emss, task)
-      emss, _emss_out_region = filter_ids_in_region(emss, "Provider")
-      assert_rbac(model, emss)
-
-      return if emss.empty?
-
-      if task == "destroy"
-        model.where(:id => emss).order("lower(name)").each do |ems|
-          id = ems.id
-          ems_name = ems.name
-          audit = {:event        => "ems_record_delete_initiated",
-                   :message      => "[#{ems_name}] Record delete initiated",
-                   :target_id    => id,
-                   :target_class => model.to_s,
-                   :userid       => session[:userid]}
-          AuditEvent.success(audit)
-        end
-        model.destroy_queue(emss)
-        add_flash(n_("Delete initiated for %{count} %{model} from the %{product} Database",
-                     "Delete initiated for %{count} %{models} from the %{product} Database", emss.length) %
-          {:count   => emss.length,
-           :product => Vmdb::Appliance.PRODUCT_NAME,
-           :model   => ui_lookup(:table => table_name),
-           :models  => ui_lookup(:tables => table_name)}) if @flash_array.nil?
-      elsif task == "pause_ems" || task == "resume_ems"
-        action = task.split("_").first
-        model.where(:id => emss).order("lower(name)").each do |ems|
-          id = ems.id
-          ems_name = ems.name
-          audit = {:event        => "ems_record_#{action}_initiated",
-                   :message      => "[#{ems_name}] Record #{action} initiated",
-                   :target_id    => id,
-                   :target_class => model.to_s,
-                   :userid       => session[:userid]}
-          AuditEvent.success(audit)
-
-          ems.pause! if action == "pause"
-          ems.resume! if action == "resume"
-        end
-      else
-        model.where(:id => emss).order("lower(name)").each do |ems|
-          id = ems.id
-          ems_name = ems.name
-          begin
-            ems.send(task.to_sym) if ems.respond_to?(task)    # Run the task
-          rescue => bang
-            add_flash(_("%{model} \"%{name}\": Error during '%{task}': %{error_message}") %
-              {:model => ui_lookup(:table => @table_name), :name => ems_name, :task => _(task.titleize), :error_message => bang.message}, :error)
-            AuditEvent.failure(:userid       => session[:userid],
-                               :event        => "#{table_name}_#{task}",
-                               :message      => "#{ems_name}: Error during '#{task}': #{bang.message}",
-                               :target_class => model.to_s, :target_id => id)
-          else
-            add_flash(_("%{model} \"%{name}\": %{task} successfully initiated") % {:model => ui_lookup(:table => @table_name), :name => ems_name, :task => _(task.titleize)})
-            AuditEvent.success(:userid       => session[:userid],
-                               :event        => "#{table_name}_#{task}",
-                               :message      => "#{ems_name}: '#{task}' successfully initiated",
-                               :target_class => model.to_s, :target_id => id)
-          end
-        end
-      end
-    end
-
     # Delete all selected or single displayed ems(s)
     def deleteemss
       assert_privileges(params[:pressed])
@@ -707,52 +648,5 @@ module Mixins
         @refresh_partial = "layouts/gtl"
       end
     end
-
-    def call_ems_pause_resume(emss, options)
-      action = if options[:resume]
-                 "resume"
-               elsif options[:pause]
-                 "pause"
-               end
-
-      process_emss(emss, "#{action}_ems") unless emss.empty?
-      return if @flash_array.present?
-      add_flash(n_("%{action} initiated for %{count} %{model} from the %{product} Database",
-                   "%{action} initiated for %{count} %{models} from the %{product} Database", emss.length) %
-                  {:count   => emss.length,
-                   :action  => action.capitalize,
-                   :product => Vmdb::Appliance.PRODUCT_NAME,
-                   :model   => ui_lookup(:table => table_name),
-                   :models  => ui_lookup(:tables => table_name)})
-    end
-
-    def pause_or_resume_emss(options)
-      assert_privileges(params[:pressed])
-      if @lastaction == "show_list"
-        emss = find_checked_items
-        if emss.empty?
-          add_flash(_("No %{model} were selected for pause") % {:model => ui_lookup(:table => table_name)}, :error)
-        end
-        call_ems_pause_resume(emss, options)
-        show_list
-        @refresh_partial = "layouts/gtl"
-      else
-        if params[:id].nil? || model.find_by_id(params[:id]).nil?
-          add_flash(_("%{record} no longer exists") % {:record => ui_lookup(:table => table_name)}, :error)
-        else
-          call_ems_pause_resume([params[:id]], options)
-        end
-        params[:display] = @display
-      end
-    end
-
-    def model
-      self.class.model
-    end
-
-    def permission_prefix
-      self.class.permission_prefix
-    end
-
   end
 end
